+++ .vscode/settings.json
@@ -13,7 +13,7 @@
     "**/CVS": true,
     "**/.DS_Store": true,
     "**/node_modules": true,
-    "**/build": true,
+    // "**/build": true,
     // Hide all spec files
     // "packages/*/src/**/*.spec.ts": true,
   },
