+++ package.json
@@ -29,24 +29,22 @@
     "@acme-ra/client": "^0.0.15",
     "@peculiar/webcrypto": "^1.4.1",
     "@peculiar/x509": "^1.9.3",
-    "@rollup/plugin-commonjs": "^22.0.0",
-    "@rollup/plugin-node-resolve": "^13.2.1",
+    "@rollup/plugin-commonjs": "^24.0.1",
+    "@rollup/plugin-node-resolve": "^15.0.1",
     "@types/mocha": "^10.0.1",
-    "@types/node": "^17.0.27",
-    "@types/node-fetch": "^2",
-    "@types/pako": "^1.0.1",
+    "@types/node": "^18.15.3",
+    "@types/pako": "^2.0.0",
     "@types/pvutils": "^1.0.0",
-    "@typescript-eslint/eslint-plugin": "^5.21.0",
-    "@typescript-eslint/parser": "^5.21.0",
-    "eslint": "^8.14.0",
-    "mocha": "^10.1.0",
-    "node-fetch": "^2.6.1",
-    "rollup": "^2.70.2",
-    "rollup-plugin-dts": "^5.1.1",
-    "rollup-plugin-typescript2": "^0.31.2",
+    "@typescript-eslint/eslint-plugin": "^5.55.0",
+    "@typescript-eslint/parser": "^5.55.0",
+    "eslint": "^8.36.0",
+    "mocha": "^10.2.0",
+    "rollup": "^3.19.1",
+    "rollup-plugin-dts": "^5.2.0",
+    "rollup-plugin-typescript2": "^0.34.1",
     "ts-node": "^10.9.1",
     "tsconfig-paths": "^4.1.2",
     "tsyringe": "^4.7.0",
-    "typescript": "^4.9.4"
+    "typescript": "^4.9.5"
   }
 }
