+++ packages/cli/rollup.config.mjs
@@ -1,12 +1,11 @@
-import path from "path";
 import resolve from "@rollup/plugin-node-resolve";
 import commonjs from "@rollup/plugin-commonjs";
 import typescript from "rollup-plugin-typescript2";
-import dts from "rollup-plugin-dts";
+import pkg from "./package.json" assert { type: "json" };
 
-const pkg = require("./package.json");
-
-const banner = [].join("\n");
+const banner = [
+  "#!/usr/bin/env node"
+].join("\n");
 const input = "src/index.ts";
 
 export default [
@@ -35,29 +34,6 @@ export default [
         banner,
         file: pkg.main,
         format: "cjs",
-      },
-      {
-        banner,
-        file: pkg.module,
-        format: "es",
-      },
-    ]
-  },
-  {
-    input,
-    external: Object.keys(pkg.dependencies),
-    plugins: [
-      dts({
-        tsconfig: path.resolve(__dirname, "./tsconfig.compile.json"),
-        compilerOptions: {
-          removeComments: false,
-        }
-      })
-    ],
-    output: [
-      {
-        banner,
-        file: pkg.types,
       }
     ]
   },
