+++ packages/copy/rollup.config.mjs
@@ -1,12 +1,14 @@
+import path from "node:path";
+import url from "node:url";
 import resolve from "@rollup/plugin-node-resolve";
 import commonjs from "@rollup/plugin-commonjs";
 import typescript from "rollup-plugin-typescript2";
+import dts from "rollup-plugin-dts";
+import pkg from "./package.json" assert { type: "json" };
 
-const pkg = require("./package.json");
-
-const banner = [
-  "#!/usr/bin/env node"
-].join("\n");
+const __filename = url.fileURLToPath(import.meta.url);
+const __dirname = path.dirname(__filename);
+const banner = [].join("\n");
 const input = "src/index.ts";
 
 export default [
@@ -35,6 +37,29 @@ export default [
         banner,
         file: pkg.main,
         format: "cjs",
+      },
+      {
+        banner,
+        file: pkg.module,
+        format: "es",
+      },
+    ]
+  },
+  {
+    input,
+    external: Object.keys(pkg.dependencies),
+    plugins: [
+      dts({
+        tsconfig: path.resolve(__dirname, "./tsconfig.compile.json"),
+        compilerOptions: {
+          removeComments: false,
+        }
+      })
+    ],
+    output: [
+      {
+        banner,
+        file: pkg.types,
       }
     ]
   },
