+++ packages/doc/package.json
@@ -31,10 +31,10 @@
     "url": "https://github.com/PeculiarVentures/pdf/issues"
   },
   "dependencies": {
-    "@peculiar/asn1-ocsp": "^2.1.4",
-    "@peculiar/asn1-schema": "^2.1.0",
-    "@peculiar/asn1-tsp": "^2.1.4",
-    "@peculiar/asn1-x509": "^2.1.4",
+    "@peculiar/asn1-ocsp": "^2.3.6",
+    "@peculiar/asn1-schema": "^2.3.6",
+    "@peculiar/asn1-tsp": "^2.3.6",
+    "@peculiar/asn1-x509": "^2.3.6",
     "@peculiar/x509": "^1.9.3",
     "@peculiarventures/pdf-copy": "^1.10.2",
     "@peculiarventures/pdf-core": "^1.10.2",
@@ -43,7 +43,7 @@
     "bytestreamjs": "2.0.1",
     "fast-png": "^6.1.0",
     "pako": "^2.0.3",
-    "pkijs": "^3.0.13",
+    "pkijs": "^3.0.14",
     "pvtsutils": "^1.2.2",
     "pvutils": "^1.1.3",
     "tslib": "^2.5.0"
