+++ packages/font/package.json
@@ -32,7 +32,7 @@
     "url": "https://github.com/PeculiarVentures/pdf/issues"
   },
   "dependencies": {
-    "@peculiar/asn1-schema": "^2.1.0",
+    "@peculiar/asn1-schema": "^2.3.6",
     "bytestreamjs": "2.0.1",
     "fontjs": "^3.0.1",
     "pako": "^2.0.4",
